console: Remove api usage in gateway-api-keys components
